pyproject.toml
[tool.poetry]
name = "wfm-schedule-updater"
version = "0.1.0"
description = ""
authors = ["Logan Dang <logandang100@gmail.com>"]
readme = "README.md"
packages = [{include = "wfm_schedule_updater"}]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.8.0"
pytz = "^2022.7.1"
python-dotenv = "^0.21.1"
gcsa = "^2.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
